+++ CHANGELOG.md
@@ -7,6 +7,8 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [0.7.1] - 2024-12-19
+
 ### Added
 
 - Zeroize memory held by `PrivateKey` when dropped.
@@ -355,7 +357,8 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - First public release.
 
-[Unreleased]: https://github.com/HMIProject/open62541/compare/v0.7.0...HEAD
+[Unreleased]: https://github.com/HMIProject/open62541/compare/v0.7.1...HEAD
+[0.7.1]: https://github.com/HMIProject/open62541/compare/v0.7.0...v0.7.1
 [0.7.0]: https://github.com/HMIProject/open62541/compare/v0.6.6...v0.7.0
 [0.6.6]: https://github.com/HMIProject/open62541/compare/v0.6.5...v0.6.6
 [0.6.5]: https://github.com/HMIProject/open62541/compare/v0.6.4...v0.6.5